Move electron config into the config folder
